venvs/api/requirements.txt
Django>=1.7.0,<1.8.0
djangorestframework>=2.3.9,<3.0.0
requests>=2.0.1,<2.5.0
South>=0.8.4,<0.9.0
